+++ src/components/friend-comp.tsx
