+++ src/components/friendsList-comp.tsx
