+++ src/components/post-comp.tsx
